Configure /projects/x and expand Sanity utils
- Remove many default styles and add some basic CSS of my own
- Render each project on the home page as a (poorly styled) card, which
links to a project page for that project when clicked
- Adds routing for `/project`, which creates a page for each project
that uses the project's slug. For now, page only displays project name.

Configurations:
- Adds handling of images from Sanity CDN to next.config, per Next docs
- Creates `sanity/config/client-config` to hold the config used to
create the Sanity client for accessing the API. This can then be used
with the `createClient(config)` function imported from `next-sanity`
anywhere Sanity data needs to be fetched (as in `sanity-utils`)
- Updates `sanity-utils` to use the helper client config
- Creates sanity-util function to fetch a single project from Sanity,
using a `slug` param